task can be removed only if it is not assigned to user

+++ database/procedures/task/RemoveTask.sql
@@ -2,8 +2,10 @@ CREATE PROCEDURE dbo.spTask_RemoveTask @Id INT
 AS
 BEGIN
     SET NOCOUNT ON
-
-    DELETE FROM task WHERE id = @Id
+    
+    -- task can be removed only if it is not assigned to user
+    IF (NOT EXISTS (SELECT * FROM result WHERE task_id = @Id))
+        DELETE FROM task WHERE id = @Id
 END
 go
